requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
chromadb
sentence-transformers
google-genai